Only show one of each upcoming event on the Events page. (#88)
This is an unrequested feature to skip all duplicate event names.
This way you only see the closest upcoming entry of any given event.
I don't know if it is a good idea or not. I don't like seeing the
duplicaduplicate
events, because I find it hard to tell which are different and which
are duplicates, but that is just me.
Another idea would be to maybe show a less detailed view for the
duplicatduplicates,
like leave out the description and make the duplicates smaller.